requirements.txt
pyxel-1.0.0